thesteady/requirements.txt
flask
requests
python-dotenv
Pillow
pytesseract
